+++ src/definitions/common.ts
@@ -1,7 +1,17 @@
 import { FormField } from '../lib/state/form'
 import { UseUnit } from '../lib/utils/units'
 
-function pushUnitFields (fields: FormField[], fieldNamePrefix: string, fieldEditNamePrefix: string, useUnitWeight: UseUnit, useUnitLength: UseUnit, useUnitSurface: UseUnit, useUnitVolume: UseUnit): void {
+function pushUnitFields (
+  fields: FormField[],
+  fieldNamePrefix: string,
+  fieldEditNamePrefix: string,
+  useUnitWeight: UseUnit,
+  useUnitLength: UseUnit,
+  useUnitWidth: UseUnit,
+  useUnitHeight: UseUnit,
+  useUnitSurface: UseUnit,
+  useUnitVolume: UseUnit
+): void {
   if (useUnitWeight !== '0') {
     fields.push({
       type: 'float',
@@ -27,6 +37,34 @@ function pushUnitFields (fields: FormField[], fieldNamePrefix: string, fieldEdit
       step: 0.01,
       edit: {
         getDataFromSourceKey: fieldEditNamePrefix + 'length'
+      }
+    })
+  }
+  if (useUnitWidth !== '0') {
+    fields.push({
+      type: 'float',
+      name: fieldNamePrefix + 'width',
+      label: 'Largeur unitaire (m)',
+      mandatory: useUnitWidth === 'mandatory',
+      min: 0,
+      max: 1000,
+      step: 0.01,
+      edit: {
+        getDataFromSourceKey: fieldEditNamePrefix + 'width'
+      }
+    })
+  }
+  if (useUnitHeight !== '0') {
+    fields.push({
+      type: 'float',
+      name: fieldNamePrefix + 'height',
+      label: 'Hauteur unitaire (m)',
+      mandatory: useUnitHeight === 'mandatory',
+      min: 0,
+      max: 1000,
+      step: 0.01,
+      edit: {
+        getDataFromSourceKey: fieldEditNamePrefix + 'height'
       }
     })
   }